+++ src/components/common/ImagePreview.jsx
@@ -1,4 +1,6 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, {
+  useCallback, useEffect, useMemo, useState
+} from 'react';
 import { createPortal } from 'react-dom';
 import styled from 'styled-components';
 import ArrowLeftIcon from 'mdi-react/ArrowLeftThickIcon';
@@ -42,7 +44,7 @@ const ImagePreviewContainer = ({ children }) => (
 
 const PreviewContainer = styled.div`
   position: fixed;
-  z-index: 4;
+  z-index: 5;
   top: 0;
   left: 0;
   height: 100%;
@@ -140,7 +142,9 @@ const ImagePreview = ({ images, startIndex = 0, onClose }) => {
   }, [index, width]);
 
   const bind = useDrag(
-    ({ active, movement: [mx], direction: [xDir], distance, cancel }) => {
+    ({
+      active, movement: [mx], direction: [xDir], distance, cancel
+    }) => {
       if (active && distance > width / 4) {
         const newIndex = clamp(
           index + (xDir > 0 ? -1 : 1),
